+++ schema.graphql
@@ -831,7 +831,7 @@ type ChangelogEntryRecord implements RecordInterface {
   _unpublishingScheduledAt: DateTime
   _updatedAt: DateTime!
   categories: [ChangelogCategoryRecord!]!
-  content: ChangelogEntryModelContentField
+  content: ChangelogEntryModelContentField!
   createdAt: DateTime!
   id: ItemId!
   seo: SeoField
