+++ src/lib/datocms/graphql-env.d.ts
@@ -47,7 +47,7 @@ export type introspection_types = {
     'ChangelogEntryModelContentField': { kind: 'OBJECT'; name: 'ChangelogEntryModelContentField'; fields: { 'blocks': { name: 'blocks'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'UNION'; name: 'ChangelogEntryModelContentBlocksField'; ofType: null; }; }; }; } }; 'links': { name: 'links'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; }; }; } }; 'value': { name: 'value'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'JsonField'; ofType: null; }; } }; }; };
     'ChangelogEntryModelFilter': { kind: 'INPUT_OBJECT'; name: 'ChangelogEntryModelFilter'; isOneOf: false; inputFields: [{ name: '_createdAt'; type: { kind: 'INPUT_OBJECT'; name: 'CreatedAtFilter'; ofType: null; }; defaultValue: null }, { name: 'createdAt'; type: { kind: 'INPUT_OBJECT'; name: 'CreatedAtFilter'; ofType: null; }; defaultValue: null }, { name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'ItemIdFilter'; ofType: null; }; defaultValue: null }, { name: '_firstPublishedAt'; type: { kind: 'INPUT_OBJECT'; name: 'PublishedAtFilter'; ofType: null; }; defaultValue: null }, { name: '_publicationScheduledAt'; type: { kind: 'INPUT_OBJECT'; name: 'PublishedAtFilter'; ofType: null; }; defaultValue: null }, { name: '_unpublishingScheduledAt'; type: { kind: 'INPUT_OBJECT'; name: 'PublishedAtFilter'; ofType: null; }; defaultValue: null }, { name: '_publishedAt'; type: { kind: 'INPUT_OBJECT'; name: 'PublishedAtFilter'; ofType: null; }; defaultValue: null }, { name: '_status'; type: { kind: 'INPUT_OBJECT'; name: 'StatusFilter'; ofType: null; }; defaultValue: null }, { name: '_updatedAt'; type: { kind: 'INPUT_OBJECT'; name: 'UpdatedAtFilter'; ofType: null; }; defaultValue: null }, { name: 'updatedAt'; type: { kind: 'INPUT_OBJECT'; name: 'UpdatedAtFilter'; ofType: null; }; defaultValue: null }, { name: '_isValid'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanFilter'; ofType: null; }; defaultValue: null }, { name: 'showInBlog'; type: { kind: 'INPUT_OBJECT'; name: 'BooleanFilter'; ofType: null; }; defaultValue: null }, { name: 'content'; type: { kind: 'INPUT_OBJECT'; name: 'StructuredTextFilter'; ofType: null; }; defaultValue: null }, { name: 'categories'; type: { kind: 'INPUT_OBJECT'; name: 'LinksFilter'; ofType: null; }; defaultValue: null }, { name: 'seo'; type: { kind: 'INPUT_OBJECT'; name: 'SeoFilter'; ofType: null; }; defaultValue: null }, { name: 'title'; type: { kind: 'INPUT_OBJECT'; name: 'StringFilter'; ofType: null; }; defaultValue: null }, { name: 'slug'; type: { kind: 'INPUT_OBJECT'; name: 'SlugFilter'; ofType: null; }; defaultValue: null }, { name: 'OR'; type: { kind: 'LIST'; name: never; ofType: { kind: 'INPUT_OBJECT'; name: 'ChangelogEntryModelFilter'; ofType: null; }; }; defaultValue: null }, { name: 'AND'; type: { kind: 'LIST'; name: never; ofType: { kind: 'INPUT_OBJECT'; name: 'ChangelogEntryModelFilter'; ofType: null; }; }; defaultValue: null }]; };
     'ChangelogEntryModelOrderBy': { name: 'ChangelogEntryModelOrderBy'; enumValues: '_createdAt_ASC' | '_createdAt_DESC' | 'createdAt_ASC' | 'createdAt_DESC' | 'id_ASC' | 'id_DESC' | '_firstPublishedAt_ASC' | '_firstPublishedAt_DESC' | '_publicationScheduledAt_ASC' | '_publicationScheduledAt_DESC' | '_unpublishingScheduledAt_ASC' | '_unpublishingScheduledAt_DESC' | '_publishedAt_ASC' | '_publishedAt_DESC' | '_status_ASC' | '_status_DESC' | '_updatedAt_ASC' | '_updatedAt_DESC' | 'updatedAt_ASC' | 'updatedAt_DESC' | '_isValid_ASC' | '_isValid_DESC' | 'showInBlog_ASC' | 'showInBlog_DESC' | 'title_ASC' | 'title_DESC'; };
-    'ChangelogEntryRecord': { kind: 'OBJECT'; name: 'ChangelogEntryRecord'; fields: { '_createdAt': { name: '_createdAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; '_editingUrl': { name: '_editingUrl'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; '_firstPublishedAt': { name: '_firstPublishedAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_isValid': { name: '_isValid'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'BooleanType'; ofType: null; }; } }; '_modelApiKey': { name: '_modelApiKey'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; '_publicationScheduledAt': { name: '_publicationScheduledAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_publishedAt': { name: '_publishedAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_seoMetaTags': { name: '_seoMetaTags'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Tag'; ofType: null; }; }; }; } }; '_status': { name: '_status'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'ENUM'; name: 'ItemStatus'; ofType: null; }; } }; '_unpublishingScheduledAt': { name: '_unpublishingScheduledAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_updatedAt': { name: '_updatedAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; 'categories': { name: 'categories'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'ChangelogCategoryRecord'; ofType: null; }; }; }; } }; 'content': { name: 'content'; type: { kind: 'OBJECT'; name: 'ChangelogEntryModelContentField'; ofType: null; } }; 'createdAt': { name: 'createdAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; 'id': { name: 'id'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'ItemId'; ofType: null; }; } }; 'seo': { name: 'seo'; type: { kind: 'OBJECT'; name: 'SeoField'; ofType: null; } }; 'showInBlog': { name: 'showInBlog'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'BooleanType'; ofType: null; }; } }; 'slug': { name: 'slug'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'title': { name: 'title'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'updatedAt': { name: 'updatedAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; }; };
+    'ChangelogEntryRecord': { kind: 'OBJECT'; name: 'ChangelogEntryRecord'; fields: { '_createdAt': { name: '_createdAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; '_editingUrl': { name: '_editingUrl'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; '_firstPublishedAt': { name: '_firstPublishedAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_isValid': { name: '_isValid'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'BooleanType'; ofType: null; }; } }; '_modelApiKey': { name: '_modelApiKey'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; '_publicationScheduledAt': { name: '_publicationScheduledAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_publishedAt': { name: '_publishedAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_seoMetaTags': { name: '_seoMetaTags'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Tag'; ofType: null; }; }; }; } }; '_status': { name: '_status'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'ENUM'; name: 'ItemStatus'; ofType: null; }; } }; '_unpublishingScheduledAt': { name: '_unpublishingScheduledAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_updatedAt': { name: '_updatedAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; 'categories': { name: 'categories'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'ChangelogCategoryRecord'; ofType: null; }; }; }; } }; 'content': { name: 'content'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'ChangelogEntryModelContentField'; ofType: null; }; } }; 'createdAt': { name: 'createdAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; 'id': { name: 'id'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'ItemId'; ofType: null; }; } }; 'seo': { name: 'seo'; type: { kind: 'OBJECT'; name: 'SeoField'; ofType: null; } }; 'showInBlog': { name: 'showInBlog'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'BooleanType'; ofType: null; }; } }; 'slug': { name: 'slug'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; 'title': { name: 'title'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; 'updatedAt': { name: 'updatedAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; }; };
     'ChangelogRecord': { kind: 'OBJECT'; name: 'ChangelogRecord'; fields: { '_createdAt': { name: '_createdAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; '_editingUrl': { name: '_editingUrl'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; '_firstPublishedAt': { name: '_firstPublishedAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_isValid': { name: '_isValid'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'BooleanType'; ofType: null; }; } }; '_modelApiKey': { name: '_modelApiKey'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; '_publicationScheduledAt': { name: '_publicationScheduledAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_publishedAt': { name: '_publishedAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_seoMetaTags': { name: '_seoMetaTags'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Tag'; ofType: null; }; }; }; } }; '_status': { name: '_status'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'ENUM'; name: 'ItemStatus'; ofType: null; }; } }; '_unpublishingScheduledAt': { name: '_unpublishingScheduledAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_updatedAt': { name: '_updatedAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; 'createdAt': { name: 'createdAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; 'id': { name: 'id'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'ItemId'; ofType: null; }; } }; 'seo': { name: 'seo'; type: { kind: 'OBJECT'; name: 'SeoField'; ofType: null; } }; 'updatedAt': { name: 'updatedAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; }; };
     'CloneButtonFormModelFilter': { kind: 'INPUT_OBJECT'; name: 'CloneButtonFormModelFilter'; isOneOf: false; inputFields: [{ name: 'id'; type: { kind: 'INPUT_OBJECT'; name: 'ItemIdFilter'; ofType: null; }; defaultValue: null }, { name: 'OR'; type: { kind: 'LIST'; name: never; ofType: { kind: 'INPUT_OBJECT'; name: 'CloneButtonFormModelFilter'; ofType: null; }; }; defaultValue: null }, { name: 'AND'; type: { kind: 'LIST'; name: never; ofType: { kind: 'INPUT_OBJECT'; name: 'CloneButtonFormModelFilter'; ofType: null; }; }; defaultValue: null }]; };
     'CloneButtonFormRecord': { kind: 'OBJECT'; name: 'CloneButtonFormRecord'; fields: { '_createdAt': { name: '_createdAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; '_editingUrl': { name: '_editingUrl'; type: { kind: 'SCALAR'; name: 'String'; ofType: null; } }; '_firstPublishedAt': { name: '_firstPublishedAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_isValid': { name: '_isValid'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'BooleanType'; ofType: null; }; } }; '_modelApiKey': { name: '_modelApiKey'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'String'; ofType: null; }; } }; '_publicationScheduledAt': { name: '_publicationScheduledAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_publishedAt': { name: '_publishedAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_seoMetaTags': { name: '_seoMetaTags'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'LIST'; name: never; ofType: { kind: 'NON_NULL'; name: never; ofType: { kind: 'OBJECT'; name: 'Tag'; ofType: null; }; }; }; } }; '_status': { name: '_status'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'ENUM'; name: 'ItemStatus'; ofType: null; }; } }; '_unpublishingScheduledAt': { name: '_unpublishingScheduledAt'; type: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; } }; '_updatedAt': { name: '_updatedAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; 'createdAt': { name: 'createdAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; 'id': { name: 'id'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'ItemId'; ofType: null; }; } }; 'updatedAt': { name: 'updatedAt'; type: { kind: 'NON_NULL'; name: never; ofType: { kind: 'SCALAR'; name: 'DateTime'; ofType: null; }; } }; }; };
